requirements.txt
# Core dependencies
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
redis==5.0.1
motor==3.3.2
pymongo==4.6.1
mongomock==4.1.2

# Machine learning
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Monitoring and logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Task queue
celery==5.3.6

# Email
python-jose==3.3.0
emails==0.6

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

# Web Framework
starlette>=0.31.0

# ML and Data Processing
hyperliquid-python-sdk>=0.1.0
python-binance>=1.0.19  # For reference implementation and testing
ccxt>=4.1.0  # For multi-exchange support
ta-lib>=0.4.28  # Technical analysis library

# Security and Authentication
fastapi-mail==1.4.1
pyotp==2.9.0
qrcode[pil]==7.4.2
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring and Metrics
grafana-api>=1.0.3
jinja2>=3.1.0  # For alert templates

# WebSockets
websockets>=11.0.3

# Plotting
plotly>=5.18.0
dash>=2.14.0
matplotlib==3.7.2

# Other
boto3>=1.28.0
requests>=2.31.0
pyyaml>=6.0.1
retrying==1.3.4
selenium==4.15.2
playwright==1.40.0
locust==2.20.0
safety==2.3.5
bandit==1.7.5